backend/requirements.txt
# FastAPI 프레임워크
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# CORS 미들웨어
python-multipart==0.0.6

# 환경 변수
python-dotenv==1.0.0

# HTTP 요청
requests==2.31.0

# MySQL
pymysql==1.1.0

# AWS S3
boto3==1.29.7

# 이미지 처리
Pillow==10.1.0
